pyproject.toml
[project]
name = "my-app"
version = "0.1.0"
description = "AI-powered full-stack application"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "pyjwt>=2.12.1",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
]
sdk = [
    "claude-agent-sdk>=0.1.54",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
